+++ workspace/solutions/kotlin/lib/src/test/kotlin/s1/s8/s9/s8/s_1898_test.kt
@@ -1,0 +1,16 @@
+package s1.s8.s9.s8
+
+import utils.Test_
+import kotlin.test.Test
+import s1.s8.s9.s8.s_1898_00.Solution as s_1898_00
+
+class s_1898_test {
+    @Test
+    fun test() {
+        val test = Test_(
+            this::class,
+            s_1898_00::class,
+        )
+        test.run()
+    }
+}