+++ workspace/solutions/rust/src/s1/s8/mod.rs
@@ -1,1 +1,2 @@
 pub mod s7;
+pub mod s9;