+++ workspace/solutions/rust/src/s1/s8/s9/s8/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_1898_00;
+pub mod s_1898_test;
+
+pub type TSolution = fn(String, String, Vec<i32>) -> i32;