+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - 添加题目 `1870` 的题解 | Add the solution to the question `1870`
   - [1870. 准时到达的列车最小时速 - 力扣（LeetCode）](https://leetcode.cn/problems/minimum-speed-to-arrive-on-time/description/)
 - 完善题目 `0912` 的题解 | Improve the solution to the question `0912`
+- 添加题目 `1898` 的题解 | Add the solution to the question `1898`
+  - [1898. 可移除字符的最大数目 - 力扣（LeetCode）](https://leetcode.cn/problems/maximum-number-of-removable-characters/description/)
 
 ## 2024-08-21